Strip utils of all other submodule dependencies
