+++ pyEpiabm/pyEpiabm/py2c/__init__.py
@@ -1,0 +1,10 @@
+#
+# Py2C subpackage of the pyEpiabm module.
+#
+
+""" pyEpiabm.py2c provides methods used to convert python
+classes to C, for speeding up simulations
+"""
+
+
+from .py2c_population import py2c_population
