+++ pyEpiabm/pyEpiabm/sweep/__init__.py
@@ -17,3 +17,4 @@ from .queue_sweep import QueueSweep
 from .spatial_sweep import SpatialSweep
 from .update_place_sweep import UpdatePlaceSweep
 from .intervention_sweep import InterventionSweep
+from .transition_matrices import StateTransitionMatrix, TransitionTimeMatrix
