+++ pyEpiabm/pyEpiabm/utility/__init__.py
@@ -12,6 +12,3 @@ from .covidsim_kernel import SpatialKernel
 from .random_methods import RandomMethods
 from .inverse_cdf import InverseCdf
 from .exception_logger import log_exceptions
-from .state_transition_matrix import StateTransitionMatrix
-from .transition_time_matrix import TransitionTimeMatrix
-from .py2c_population import py2c_population
